+++ integration_tests/test_decorators.py
@@ -1,5 +1,6 @@
 from threading import Thread
 from time import sleep
+from typing import Any
 
 import pytest
 
@@ -8,7 +9,7 @@ from py_cachify import CachifyLockError, cached, once
 
 def test_once_decorator():
     @once(key='test_key-{arg1}', return_on_locked='IF_LOCKED')
-    def _sync_function(arg1, arg2):
+    def _sync_function(arg1: int, arg2: int):
         sleep(2)
         return arg1 + arg2
 
@@ -22,7 +23,7 @@ def test_once_decorator():
 
 def test_once_decorator_raises():
     @once(key='test_key-{arg1}-{arg2}', raise_on_locked=True)
-    def _sync_function(arg1, arg2):
+    def _sync_function(arg1: int, arg2: int):
         sleep(2)
         return arg1 + arg2
 
@@ -36,7 +37,7 @@ def test_once_decorator_raises():
 @pytest.mark.asyncio
 async def test_once_decorator_async_function():
     @once(key='async_test_key-{arg1}-{arg2}', return_on_locked='IF_LOCKED')
-    async def _async_function(arg1, arg2, initial=True):
+    async def _async_function(arg1: int, arg2: int, initial: bool = True) -> Any:
         res = None
         if initial:
             res = await _async_function(arg1, arg2, initial=False)
@@ -48,7 +49,7 @@ async def test_once_decorator_async_function():
 
 
 @pytest.mark.asyncio
-async def test_async_once_decorator_raise_on_locked(init_cachify_fixture):
+async def test_async_once_decorator_raise_on_locked(init_cachify_fixture: Any):
     @once(key='async_test_key-{arg1}-{arg2}', raise_on_locked=True)
     async def _async_function(arg1: int, arg2: int) -> int:
         await _async_function(arg1, arg2)
@@ -60,7 +61,7 @@ async def test_async_once_decorator_raise_on_locked(init_cachify_fixture):
 
 def test_cached_decorator_sync_function():
     @cached(key='test_key')
-    def _sync_function_wrapped(arg1, arg2):
+    def _sync_function_wrapped(arg1: int, arg2: int):
         return arg1 + arg2
 
     result = _sync_function_wrapped(3, 4)
@@ -76,7 +77,7 @@ def test_cached_decorator_sync_function():
 @pytest.mark.asyncio
 async def test_cached_decorator_async_function():
     @cached(key='test_key_{arg1}')
-    async def _async_function_wrapped(arg1, arg2):
+    async def _async_function_wrapped(arg1: int, arg2: int):
         return arg1 + arg2
 
     result = await _async_function_wrapped(3, 4)