+++ tests/conftest.py
@@ -1,3 +1,6 @@
+# pyright: reportPrivateUsage=false
+from typing import Any
+
 import pytest
 
 import py_cachify._backend._lib
@@ -5,8 +8,9 @@ from py_cachify import init_cachify
 
 
 @pytest.fixture(scope='function')
-def init_cachify_fixture():
+def init_cachify_fixture() -> Any:
     init_cachify()
     yield
-    py_cachify._backend._lib._cachify._sync_client._cache = {}
+    assert py_cachify._backend._lib._cachify
+    py_cachify._backend._lib._cachify._sync_client._cache = {}  # pyright: ignore[reportAttributeAccessIssue]
     py_cachify._backend._lib._cachify = None